// sdk/tour-operations/models/pagination-response-of-tour-infor-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WebAPI - Tour Operation
 * Clean Architecture Tour Operation for .NET 7 WebApi built with Multitenancy Support.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { TourInforDto } from './tour-infor-dto';

/**
 * 
 * @export
 * @interface PaginationResponseOfTourInforDto
 */
export interface PaginationResponseOfTourInforDto {
    /**
     * 
     * @type {Array<TourInforDto>}
     * @memberof PaginationResponseOfTourInforDto
     */
    'data'?: Array<TourInforDto>;
    /**
     * 
     * @type {number}
     * @memberof PaginationResponseOfTourInforDto
     */
    'currentPage'?: number;
    /**
     * 
     * @type {number}
     * @memberof PaginationResponseOfTourInforDto
     */
    'totalPages'?: number;
    /**
     * 
     * @type {number}
     * @memberof PaginationResponseOfTourInforDto
     */
    'totalCount'?: number;
    /**
     * 
     * @type {number}
     * @memberof PaginationResponseOfTourInforDto
     */
    'pageSize'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PaginationResponseOfTourInforDto
     */
    'hasPreviousPage'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PaginationResponseOfTourInforDto
     */
    'hasNextPage'?: boolean;
}