// sdk/tour-operations/models/pagination-response-of-tour-type-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WebAPI - Tour Operation
 * Clean Architecture Tour Operation for .NET 7 WebApi built with Multitenancy Support.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { TourTypeDto } from './tour-type-dto';

/**
 * 
 * @export
 * @interface PaginationResponseOfTourTypeDto
 */
export interface PaginationResponseOfTourTypeDto {
    /**
     * 
     * @type {Array<TourTypeDto>}
     * @memberof PaginationResponseOfTourTypeDto
     */
    'data'?: Array<TourTypeDto>;
    /**
     * 
     * @type {number}
     * @memberof PaginationResponseOfTourTypeDto
     */
    'currentPage'?: number;
    /**
     * 
     * @type {number}
     * @memberof PaginationResponseOfTourTypeDto
     */
    'totalPages'?: number;
    /**
     * 
     * @type {number}
     * @memberof PaginationResponseOfTourTypeDto
     */
    'totalCount'?: number;
    /**
     * 
     * @type {number}
     * @memberof PaginationResponseOfTourTypeDto
     */
    'pageSize'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PaginationResponseOfTourTypeDto
     */
    'hasPreviousPage'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PaginationResponseOfTourTypeDto
     */
    'hasNextPage'?: boolean;
}